Pipelines etl auto/FINAL. DAG contaminacion/requirements.txt
google-cloud-storage==2.14.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==1.5.3
geopy==2.4.1
gdown==4.7.1
numpy==1.22.3